CHANGELOG + 1. Save the game state when the player dies, when the quit option is selected and when a new game is made 2. Enemies now have 3 health whereas the player has 15 and each attack from either characters will deal 1 damage to the victim
-

#
1. Optimize code
2. Google play services (online high score rankings)
