+++ core/src/com/badlogic/gamescreentest/headsUP/Map/MapHUD.java
@@ -1,24 +1,24 @@
-package com.badlogic.gamescreentest.headsUP;
+package com.badlogic.gamescreentest.headsUP.Map;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 
 /**
- * Created by admin on 7/27/2016.
+ * Created by admin on 7/28/2016.
  */
-public class gameOptions extends Actor {
+public class MapHUD extends Actor {
     Texture img;
     float x;
     float y;
     float width;
     float height;
 
-    public gameOptions(Texture texture, float cellWidth, float cellHeight) {
+    public MapHUD(Texture texture, float cellWidth, float cellHeight) {
         img = texture;
         width = cellWidth;
         height = cellHeight;
-        x = width * 9;
+        x = width * 8;
         y = height * 7;
     }
