+++ core/src/com/badlogic/gamescreentest/headsUP/Options/gameOptions.java
@@ -1,24 +1,24 @@
-package com.badlogic.gamescreentest.headsUP;
+package com.badlogic.gamescreentest.headsUP.Options;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 
 /**
- * Created by admin on 8/16/2016.
+ * Created by admin on 7/27/2016.
  */
-public class MainMenuButton extends Actor {
+public class gameOptions extends Actor {
     Texture img;
     float x;
     float y;
     float width;
     float height;
 
-    public MainMenuButton(Texture texture, float cellWidth, float cellHeight) {
+    public gameOptions(Texture texture, float cellWidth, float cellHeight) {
         img = texture;
         width = cellWidth;
         height = cellHeight;
-        x = width * 7;
+        x = width * 9;
         y = height * 7;
     }
