+++ core/src/com/badlogic/gamescreentest/headsUP/PlayerHealthBar/HPBarStyle.java
@@ -1,4 +1,4 @@
-package com.badlogic.gamescreentest;
+package com.badlogic.gamescreentest.headsUP.PlayerHealthBar;
 
 import com.badlogic.gdx.scenes.scene2d.ui.ProgressBar;
 import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
